+++ scripts/build.sh
@@ -145,5 +145,13 @@ if [ "$BUILD" = "production" ]; then
     echo -e "$LINECLEAR$BASENAME$CHECK${DIM}Output files stamped${RESET}"
 
 
+    # Copy root files (README, LICENSE, etc.)
+    printf "$BASENAME${DOT}Copying root files"
+    cp README.md dist/
+    cp LICENSE.txt dist/
+    cp CHANGELOG.md dist/
+    cp SECURITY.md dist/
+
+    node ./scripts/make-publish-package.json.js
 
 fi