+++ src/atoms/enclose.ts
@@ -157,7 +157,7 @@ export class EncloseAtom extends Atom {
 
     const borderWidth = borderDim(this.borderStyle);
 
-    // Calculate the padding
+    // Calculate the padding on each side
     const padding = context.toEm(
       !this.padding || this.padding === 'auto'
         ? { register: 'fboxsep' }
@@ -165,15 +165,18 @@ export class EncloseAtom extends Atom {
     );
     base.setStyle('position', 'relative');
     base.setStyle('display', 'inline-block');
-    base.setStyle('top', padding, 'em');
+    // base.setStyle(
+    //   'top',
+    //   `calc(${borderWidth} / 2 + ${padding - base.height}, 'em')`
+    // );
     base.setStyle('height', base.height + base.depth, 'em');
     base.setStyle('width', base.width, 'em');
-
-    const notation = new Box(null, { classes: 'ML__notation' });
 
     // The 'ML__notation' class is required to prevent the box from being
     // omitted during rendering (otherwise it would look like an empty, no-op
     // box)
+    const notation = new Box(null, { classes: 'ML__notation' });
+
     let h = base.height + base.depth + 2 * padding;
     const w = base.width + 2 * padding;
 
@@ -279,14 +282,22 @@ export class EncloseAtom extends Atom {
       }" fill="none"`;
       svg += '/>';
     }
-    notation.width = base.width + 2 * padding + wDelta;
+    // notation.width = base.width + 2 * padding + wDelta;
     notation.height = base.height + padding;
     notation.depth = base.depth + padding;
     notation.setStyle('box-sizing', 'border-box');
-    notation.setStyle('left', `calc(-${borderWidth} / 2 )`);
-    notation.setStyle('height', `${Math.floor(100 * h) / 100}em`);
-    notation.setStyle('top', `calc(${borderWidth} / 2 )`);
-    // notation.setStyle('width', `${Math.floor(100 * w) / 100}em`);
+    notation.setStyle('left', `calc(-1 * ${borderWidth} - ${padding}em)`);
+    // notation.setStyle('height', `${Math.floor(100 * h) / 100}em`);
+    notation.setStyle(
+      'height',
+      `calc( ${base.height + base.depth + 2 * padding}em + ${borderWidth})`
+    );
+    // );
+    notation.setStyle('top', `calc(-${padding}em )`);
+    notation.setStyle(
+      'width',
+      `calc(${base.width + 2 * padding}em + ${borderWidth})`
+    );
 
     if (this.backgroundcolor)
       notation.setStyle('background-color', this.backgroundcolor);
@@ -345,10 +356,10 @@ export class EncloseAtom extends Atom {
     // Set its position as relative so that the notation can be absolute
     // positioned over the base
     result.setStyle('position', 'relative');
-    result.setStyle('vertical-align', padding, 'em');
+    // result.setStyle('vertical-align', padding, 'em');
     result.setStyle(
       'height',
-      `${Math.floor(100 * (base.height + base.depth + 2 * padding)) / 100}em`
+      `${Math.floor(100 * (notation.depth + notation.height)) / 100}em`
     );
     // We set the padding later with `left` and `right` so subtract it now
     // result.setStyle('width', `${Math.floor(100 * base.width) / 100}em`);
@@ -357,7 +368,7 @@ export class EncloseAtom extends Atom {
 
     result.height = notation.height;
     result.depth = notation.depth;
-    result.width = notation.width - 2 * padding;
+    // result.width = notation.width - 2 * padding;
     result.left = padding;
     result.right = padding;