+++ src/formats/atom-to-math-ml.ts
@@ -8,6 +8,7 @@ import { BoxAtom } from 'atoms/box';
 import { OverunderAtom } from 'atoms/overunder';
 import { GenfracAtom } from 'atoms/genfrac';
 import { AccentAtom } from 'atoms/accent';
+import { LatexValue } from 'public/core-types';
 
 type MathMLStream = {
   atoms: Atom[];
@@ -736,10 +737,7 @@ function atomToMathML(atom: Atom, options: { generateID?: boolean }): string {
       ) {
         result += '<mrow>';
         if (arrayAtom.leftDelim && arrayAtom.leftDelim !== '.') {
-          result +=
-            '<mo>' +
-            (SPECIAL_DELIMS[arrayAtom.leftDelim] || arrayAtom.leftDelim) +
-            '</mo>';
+          result += `<mo>${SPECIAL_DELIMS[arrayAtom.leftDelim] || arrayAtom.leftDelim}</mo>`;
         }
       }
 
@@ -775,10 +773,7 @@ function atomToMathML(atom: Atom, options: { generateID?: boolean }): string {
         (arrayAtom.rightDelim && arrayAtom.rightDelim !== '.')
       ) {
         if (arrayAtom.rightDelim && arrayAtom.rightDelim !== '.') {
-          result +=
-            '<mo>' +
-            (SPECIAL_DELIMS[arrayAtom.leftDelim!] || arrayAtom.rightDelim) +
-            '</mo>';
+          result += `'<mo>${SPECIAL_DELIMS[arrayAtom.rightDelim!] || arrayAtom.rightDelim}</mo>`;
         }
 
         result += '</mrow>';
@@ -961,10 +956,21 @@ function atomToMathML(atom: Atom, options: { generateID?: boolean }): string {
             ';';
         } else if (typeof atom.value === 'string')
           result = atom.value.charAt(0);
-        else {
-          console.error('Did not expect this');
-          result = '';
+      } else if (atom.command === '\\char') {
+        // This is a \char command, which is a Unicode character
+        const val = atom.args?.[0] as LatexValue;
+        if (val !== undefined && 'number' in val) {
+          const codepoint = val.number;
+          if (typeof codepoint === 'number') {
+            result =
+              '&#x' + ('000000' + codepoint.toString(16)).slice(-4) + ';';
+          }
         }
+      } else if (typeof atom.value === 'string') {
+        result = atom.value;
+      } else {
+        console.error('Did not expect this');
+        result = '';
       }
 
       const tag = /\d/.test(result) ? 'mn' : 'mi';