+++ src/public/options.ts
@@ -311,7 +311,6 @@ export type EditingOptions = {
   mathVirtualKeyboardPolicy: 'auto' | 'manual' | 'sandboxed';
 };
 
-/** @internal */
 export type LayoutOptions = {
   defaultMode: 'inline-math' | 'math' | 'text';
 
@@ -480,8 +479,8 @@ export type StaticRenderOptions = Partial<LayoutOptions> & {
 
   asciiMath?: {
     delimiters?: {
-      display?: string[]; // [openDelim: string, closeDelim: string][];
-      inline?: string[]; // [openDelim: string, closeDelim: string][];
+      display?: [openDelim: string, closeDelim: string][];
+      inline?: [openDelim: string, closeDelim: string][];
     };
   };
 
@@ -499,7 +498,7 @@ export type StaticRenderOptions = Partial<LayoutOptions> & {
      * Delimiter pairs that will trigger a render of the content in
      * display style or inline, respectively.
      *
-     * **Default**: `{display: [ ['$$', '$$'], ['\\[', '\\]'] ] ], inline: [ ['\\(','\\)'] ] ]}`
+     * **Default**: `{display: [ ['$$', '$$'], ['\\[', '\\]'] ], inline: [ ['\\(','\\)'] ]}`
      *
      */
     delimiters?: {